user can change their own tickets with nice() cal
Made some changes in do_nice to reassign the new ticket value to the new
process.
schedulenice.c is actually /usr/src/servers/pm/schedule.c and the only
thing i changed here is to pass the nice value as specified by the user
to do_nice()
Make sure to rename that file when you put it in /usr/src/servers/pm!
This schedule.c file is the one in /usr/src/servers/sched